+++ packages/plugin-gallery/web/src/components/gallery-controls/gallery-image-settings.js
@@ -86,6 +86,7 @@ class ImageSettings extends Component {
     const { showNewTabCheckbox, showNoFollowCheckbox, showSponsoredCheckbox, placeholder } =
       linkPanel || {};
     const { metadata = {} } = image || {};
+    const useNewSettingsUi = experiments?.newSettingsUi?.enabled;
 
     const altText = typeof metadata.altText === 'string' ? metadata.altText : metadata.title;
     /* eslint-disable jsx-a11y/no-noninteractive-element-interactions, jsx-a11y/click-events-have-key-events */
@@ -98,9 +99,10 @@ class ImageSettings extends Component {
               onCancel={onCancel}
               onSave={onSave}
               t={t}
-              title={experiments?.newSettingsUi?.enabled && t('GallerySettings_Header')}
+              title={useNewSettingsUi && t('GallerySettings_Header')}
+              useNewSettingsUi={useNewSettingsUi}
             />
-          ) : experiments?.newSettingsUi?.enabled ? (
+          ) : useNewSettingsUi ? (
             <SettingsPanelHeader showCloseIcon={false}>
               <div onClick={onCancel}>
                 <BackIcon className={styles.galleryImageSettings_backIcon} />